XIOS : update regarding the XIOS xml xml cpp-processing
Now the cpp-processing is done outside the jobcomp.
It is done using the script process_xios_xml.sh scripts in croco/XIOS

+++ CVTK/common/jobcomp_rvtk.bash
@@ -109,10 +109,10 @@ ls ${SOURCE}/jobcomp           > /dev/null  2>&1 && \cp ${SOURCE}/jobcomp $SCRDI
 ls ${SOURCE}/amr.in            > /dev/null  2>&1 && \cp ${SOURCE}/amr.in $SCRDIR
 ls ${AGRIF_SRC}                > /dev/null  2>&1 && \cp -r ${AGRIF_SRC} $SCRDIR
 ls ${ROOT_DIR}/XIOS/*.F        > /dev/null  2>&1 && \cp ${ROOT_DIR}/XIOS/*.F $SCRDIR
-ls ${ROOT_DIR}/PISCES/*        > /dev/null  2>&1 && \cp ${ROOT_DIR}/PISCES/* $SCRDIR
+ls ${ROOT_DIR}/PISCES/*        > /dev/null  2>&1 && \cp -r ${ROOT_DIR}/PISCES/* $SCRDIR
 ls ${ROOT_DIR}/PISCES/SED/*    > /dev/null  2>&1 && \cp ${ROOT_DIR}/PISCES/SED/* $SCRDIR
 ls ${ROOT_DIR}/PISCES/kRGB61*  > /dev/null  2>&1 && \cp ${ROOT_DIR}/PISCES/kRGB61* $RUNDIR
-ls ${ROOT_DIR}/MUSTANG/*       > /dev/null  2>&1 && \cp ${ROOT_DIR}/MUSTANG/* $SCRDIR
+ls ${ROOT_DIR}/MUSTANG/*       > /dev/null  2>&1 && \cp -r ${ROOT_DIR}/MUSTANG/* $SCRDIR
 
 if [[ -e "namelist_pisces_ref" ]] ; then
         echo "  file namelist_pisces exists in Run directory"
@@ -239,11 +239,6 @@ if $($CPP1 testkeys.F | grep -i -q xiosisdefined) ; then
         FFLAGS1="$FFLAGS1 -I$XIOS_ROOT_DIR/inc"
 	
         ln -fs $XIOS_ROOT_DIR/bin/xios_server.exe $RUNDIR/.
-	#
-        $CPP1 -P -traditional -imacros cppdefs.h  ${ROOT_DIR}/XIOS/field_def.xml_full $RUNDIR/field_def.xml
-        $CPP1 -P -traditional -imacros cppdefs.h  ${ROOT_DIR}/XIOS/domain_def.xml $RUNDIR/domain_def.xml
-        #$CPP1 -P -traditional -imacros cppdefs.h  ${ROOT_DIR}/XIOS/file_def.xml $RUNDIR/file_def.xml
-        $CPP1 -P -traditional -imacros cppdefs.h  ${ROOT_DIR}/XIOS/iodef.xml $RUNDIR/iodef.xml
 fi
 
 #
@@ -337,11 +332,7 @@ rm -f flags.tmp
 $MAKE depend
 $MAKE
   
-if [ -f croco  ]; then
-    mv croco $RUNDIR
-else
-    exit 1
-fi
+[[ -f croco  ]] && mv croco $RUNDIR
 #[[ -f partit ]] && mv partit $RUNDIR
 #[[ -f ncjoin ]] && mv ncjoin  $RUNDIR
 #
